requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic-settings==2.3.0
celery==5.3.6
psycopg[binary]==3.1.18
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
httpx==0.27.0